+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [0.10.1] - (2024-01-07)
+
 ### Changed
 
 - improve OSA implementation
@@ -195,7 +197,8 @@ vector of results (thanks @ovarene)
 
 - Implement Hamming, Jaro, Jaro-Winkler, and Levenshtein
 
-[Unreleased]: https://github.com/rapidfuzz/strsim-rs/compare/0.10.0...HEAD
+[Unreleased]: https://github.com/rapidfuzz/strsim-rs/compare/0.10.1...HEAD
+[0.10.1]: https://github.com/rapidfuzz/strsim-rs/compare/0.10.0...0.10.1
 [0.10.0]: https://github.com/rapidfuzz/strsim-rs/compare/0.9.3...0.10.0
 [0.9.3]: https://github.com/rapidfuzz/strsim-rs/compare/0.9.2...0.9.3
 [0.9.2]: https://github.com/rapidfuzz/strsim-rs/compare/0.9.1...0.9.2